alternative pandas2df implementation

+++ R/utils.R
@@ -543,6 +543,13 @@ pandas2df <- function(x, use_arrow=TRUE) {
   }
   # remove index to keep arrow happy
   x$reset_index(drop=T, inplace=T)
+  if(FALSE) {
+    # in future we might prefer this, but for now let's just leave it latent
+    pa=reticulate::import('pyarrow')
+    at=pa$Table$from_pandas(x)
+    return(as.data.frame(at))
+  }
+
   tf=tempfile(fileext = '.feather')
   on.exit(unlink(tf))
   if(isTRUE(pyarrow_version()>='0.17.0') && pandas_version()>='1.1.0' ) {
